ANSAK-string Version 1.2 complete
String scanning can now be conditioned looking for assigned Unicode 8.0 values, control characters, white space characters and private-assignable characters in any positive and negative combination.

Unit testing for assigned-ness in the Unicode standard is implemented using the string-split mechanism also in ANSAK-string. If you can compile and use ANSAK-string, you don't need to have AWK on hand as well (not available on all platforms.

I have changed the README.md to express flexibility of licensing for jurisdictions that may need it, and to mention that I am open to providing commercial support or requests for conversion of other Windows Code Pages and/or ISO/ECMA 8859 standards. I have also responded to various comments on typos and spacing around angle brackets, and added some static_asserts around the types used to instantiate templates within the library, and some unit tests that were marked as TODOs in 1.1 and before.

I have maintained cross-platform compatibility with Win32 (Visual Studio 2013 or later for the library, 2015 or later for the unit tests)

CMake compatibility, down to building the test

* Added the FindCPPUnit module from Kitware's Bulletin Board (for use in Windows)
* Added code the CMakeLists.txt to handle differences between Windows and Linux as smoothly as possible