wchar_t and TCHAR support (#6)
Implements isXxxxx and toXxxxx involving wchar_t- and TCHAR-based data types
* regardless of 2-byte or 4-byte wchar_t
* regardless of 1-byte or 2-byte TCHAR
* regardless of conversion from or to wchar_t- or TCHAR- based types.

Size of wchar_t is determined by the preprocessor, based on the value of WCHAR_MAX.

Since wchar_t is used extensively on Windows for UTF-16 values, conversions to wchar_t-based types are limited to those that are compatible with UTF-16. Conversion to narrow TCHAR-based types is limited to ASCII values.